requirements.txt
amqp==5.1.1
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.3.0
blinker==1.6.2
celery==5.2.7
certifi==2023.5.7
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==4.2.1
django-bootstrap4==23.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-jet-reboot==1.3.3
django-plotly-dash==2.2.1
django-tables2==2.5.3
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
dpd-components==0.1.0
drf-yasg==1.21.5
eventlet==0.33.3
Flask==2.3.2
greenlet==2.0.2
humanize==4.6.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.25.0
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.15.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
PyJWT==2.7.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tenacity==8.2.2
tomli==2.0.1
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
